Update Program.cs
-Adding using System.Numerics;
-By BigInteger you can calculate 1000!, 10000! and even 100000! It will take some time, but OverflowException will not occur. The BigInteger class is very powerful but it works many times slower than int and long. For our unpleasant surprise there is no class "big decimal" in .NET Framework, only "big integer".

+++ Chapter6Exercise/Program.cs
@@ -5,6 +5,7 @@ using System.Threading;
 using System.Globalization;
 using System.Linq;
 using System.Collections.Generic;
+using System.Numerics;
 
 namespace Chapter6Exercises
 {
@@ -19,11 +20,11 @@ namespace Chapter6Exercises
 
             // Factorial Finder Program with do-while loop
 
-            /*
+            
             Console.Write("n = ");
             int number = int.Parse(Console.ReadLine());
 
-            decimal factorial = 1;
+            BigInteger factorial = 1;
 
             do
             {
@@ -35,7 +36,7 @@ namespace Chapter6Exercises
                 number--;
             } while (number > 0);
             Console.WriteLine("number! = " + factorial);
-            */
+            
 
             // Factorial Finder Program with while loop
             /*
